fixed all that was giving errors, kept the new file names and merged the codes.
