+++ client/src/pages/Front/index.js
@@ -3,7 +3,7 @@ import API from '../../utils/API';
 import { Link } from 'react-router-dom';
 import { Footer } from '../../components/Footer';
 import { Description } from '../../components/Description'
-import { AdminLogin } from '../../components/Modal';
+import AdminLogin from '../../components/Modal';
 import { AdminList } from '../../components/AdminList';
 
 class FrontPage extends Component {
@@ -36,7 +36,6 @@ class FrontPage extends Component {
       .then(res => this.setState({ dinnerItems: res.data[0].items }))
       .catch(err => console.log(err))
   }
-
 
   render() {
     return (
